requirements.txt
python-dotenv
langchain_groq
youtube_transcript_api
langgraph
streamlit